refactor: rename network_stats to get_dag_info
